api/serializers/service_serializer.py
from rest_framework import serializers
from api.models import Service

class ServiceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Service
        fields = ["id", "code", "label", "price"]