requirements.txt
amqp==5.3.1
asgiref==3.8.1
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
boto3==1.38.5
botocore==1.38.5
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.7
django-background-tasks==1.2.8
django-cors-headers==4.7.0
django-extensions==3.2.3
django-sslserver==0.22
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.3
MarkupSafe==3.0.2
MouseInfo==0.1.3
pdfkit==1.0.0
phonenumbers==9.0.2
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyOpenSSL==25.0.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
rubicon-objc==0.5.0
s3transfer==0.12.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wikipedia==1.4.0
